Modularize local copy overrides and streamline cross builds
