+++ .github/workflows/ci.yml
@@ -235,13 +235,9 @@ jobs:
       - uses: actions-rust-lang/setup-rust-toolchain@v1
         with:
           toolchain: 1.87
+          components: rust-src
           targets: ${{ matrix.platform.target }}
           cache: true
-
-      - name: Install Rust target support
-        run: |
-          rustup target add ${{ matrix.platform.target }}
-          rustup component add rust-src
 
       - uses: Swatinem/rust-cache@v2
         with:
